requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pyyaml>=6.0

# testing
pytest>=9.0
pytest-cov>=7.0
pytest-asyncio>=1.3
httpx>=0.28
